+++ src-tauri/src/plugins/backup.rs
@@ -1,5 +1,6 @@
-use super::fs_extra::preview_path;
+use super::fs_extra::{get_file_name, preview_path};
 use flate2::{read::GzDecoder, write::GzEncoder, Compression};
+use fs_extra::dir::{move_dir, CopyOptions};
 use std::{
     fs::{read_dir, File},
     path::PathBuf,
@@ -9,11 +10,11 @@ use tauri::{
     api::path::download_dir,
     command, generate_handler,
     plugin::{Builder, TauriPlugin},
-    Result, Wry,
+    Wry,
 };
 
 #[command]
-async fn export_data(src_dir: PathBuf, file_name: String) -> Result<()> {
+async fn export_data(src_dir: PathBuf, file_name: String) -> tauri::Result<()> {
     let dst_path = download_dir().unwrap().join(file_name.clone());
     let dst_file = File::create(dst_path.clone())?;
 
@@ -41,7 +42,7 @@ async fn export_data(src_dir: PathBuf, file_name: String) -> Result<()> {
 }
 
 #[command]
-async fn import_data(dst_dir: PathBuf, path: String) -> Result<bool> {
+async fn import_data(dst_dir: PathBuf, path: String) -> tauri::Result<bool> {
     let file = File::open(path)?;
     let decoder = GzDecoder::new(file);
     let mut archive = Archive::new(decoder);
@@ -59,8 +60,17 @@ async fn import_data(dst_dir: PathBuf, path: String) -> Result<bool> {
     Ok(true)
 }
 
+#[command]
+async fn move_data(from: PathBuf, to: PathBuf) -> Result<String, String> {
+    let options = CopyOptions::new();
+
+    move_dir(&from, &to, &options).map_err(|err| err.to_string())?;
+
+    Ok(get_file_name(from))
+}
+
 pub fn init() -> TauriPlugin<Wry> {
     Builder::new("backup")
-        .invoke_handler(generate_handler![export_data, import_data])
+        .invoke_handler(generate_handler![export_data, import_data, move_data])
         .build()
 }
