+++ src/constants/index.ts
@@ -10,7 +10,7 @@ export const LISTEN_KEY = {
 	CLEAR_HISTORY: "clear-history",
 	UPDATE: "update",
 	TRAY_CLICK: "tray-click",
-	IMPORT_DATA: "import-data",
+	CHANGE_DATA: "change-data",
 	CHANGE_LANGUAGE: "change-language",
 	TOGGLE_LISTENING: "toggle-listening",
 	SHOW_WINDOW: "show-window",
@@ -68,6 +68,7 @@ export const THEME_PLUGIN = {
 export const BACKUP_PLUGIN = {
 	EXPORT_DATA: "plugin:backup|export_data",
 	IMPORT_DATA: "plugin:backup|import_data",
+	MOVE_DATA: "plugin:backup|move_data",
 };
 
 export const LANGUAGE = {
