+++ src/pages/Backup/components/Manual/index.tsx
@@ -1,0 +1,7 @@
+import ProList from "@/components/ProList";
+
+const Manual = () => {
+	return <ProList header="导入和导出">Manual </ProList>;
+};
+
+export default Manual;
