+++ src/pages/Clipboard/Panel/index.tsx
@@ -100,7 +100,7 @@ const ClipboardPanel = () => {
 			getClipboardList();
 		});
 
-		listen(LISTEN_KEY.IMPORT_DATA, async () => {
+		listen(LISTEN_KEY.CHANGE_DATA, async () => {
 			await initDatabase();
 
 			getClipboardList();
